[#195] Move TypeDetails into separate file
